API_BACKEND/python-logic-service/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.23.0
python-dateutil>=2.8.2
pytest>=7.4.0
httpx>=0.24.1
